requirements.txt
alembic==1.0.10
aniso8601==6.0.0
app==0.0.1
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
mysql-connector==2.2.9
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4
